move YamlRootRuleConfigurations to common module
