+++ sharding-core/sharding-core-common/src/main/java/org/apache/shardingsphere/sharding/core/yaml/constructor/YamlRootRuleConfigurationsConstructor.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.sharding.core.yaml.constructor;
 
-import org.apache.shardingsphere.sharding.core.yaml.config.YamlRootRuleConfigurations;
+import org.apache.shardingsphere.underlying.common.yaml.config.YamlRootRuleConfigurations;
 import org.apache.shardingsphere.sharding.core.yaml.config.strategy.YamlNoneShardingStrategyConfiguration;
 import org.apache.shardingsphere.sharding.core.yaml.constructor.construct.YamlNoneShardingStrategyConfigurationConstruct;
 import org.apache.shardingsphere.underlying.common.yaml.constructor.AbstractTypeConstructor;
