+++ sharding-core/sharding-core-common/src/test/java/org/apache/shardingsphere/sharding/core/yaml/engine/YamlEngineRootConfigurationsTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.sharding.core.yaml.engine;
 
-import org.apache.shardingsphere.sharding.core.yaml.config.YamlRootRuleConfigurations;
+import org.apache.shardingsphere.underlying.common.yaml.config.YamlRootRuleConfigurations;
 import org.apache.shardingsphere.masterslave.core.yaml.config.YamlMasterSlaveRuleConfiguration;
 import org.apache.shardingsphere.sharding.core.yaml.config.YamlShardingRuleConfiguration;
 import org.apache.shardingsphere.sharding.core.yaml.constructor.YamlRootRuleConfigurationsConstructor;
