+++ sharding-jdbc/sharding-jdbc-orchestration/src/main/java/org/apache/shardingsphere/shardingjdbc/orchestration/internal/yaml/YamlOrchestrationRootRuleConfigurations.java
@@ -20,7 +20,7 @@ package org.apache.shardingsphere.shardingjdbc.orchestration.internal.yaml;
 import lombok.Getter;
 import lombok.Setter;
 import org.apache.shardingsphere.orchestration.center.yaml.config.YamlCenterRepositoryConfiguration;
-import org.apache.shardingsphere.sharding.core.yaml.config.YamlRootRuleConfigurations;
+import org.apache.shardingsphere.underlying.common.yaml.config.YamlRootRuleConfigurations;
 
 import java.util.Map;
